p3_matrix::bitrev

Type Alias BitReversedMatrixView

pub type BitReversedMatrixView<Inner> = RowIndexMappedView<BitReversalPerm, Inner>;

A matrix view that reorders its rows using bit-reversal.

This type is produced by applying BitReversibleMatrix::bit_reverse_rows() to a DenseMatrix.

Aliased Type

pub struct BitReversedMatrixView<Inner> {
    pub index_map: BitReversalPerm,
    pub inner: Inner,
}

Fields
index_map: BitReversalPerm

A row index mapping that defines the number and order of visible rows.

inner: Inner

The inner matrix that holds actual data.