p3_matrix::dense

Struct DenseMatrix

pub struct DenseMatrix<T, V = Vec<T>> {
    pub values: V,
    pub width: usize,
    /* private fields */
}

A dense matrix in row-major format, with customizable backing storage.

The data is stored as a flat buffer, where rows are laid out consecutively.

Fields
values: V

Flat buffer of matrix values in row-major order.

width: usize

Number of columns in the matrix.

The number of rows is implicitly determined as values.len() / width.

Implementations

impl<T: Clone + Send + Sync + Default> DenseMatrix<T>

pub fn default(width: usize, height: usize) -> Self

Create a new dense matrix of the given dimensions, backed by a Vec, and filled with default values.

impl<T: Clone + Send + Sync, S: DenseStorage<T>> DenseMatrix<T, S>

pub fn new(values: S, width: usize) -> Self

Create a new dense matrix of the given dimensions, backed by the given storage.

Note that it is undefined behavior to create a matrix such that values.len() % width != 0.

pub fn new_row(values: S) -> Self

Create a new RowMajorMatrix containing a single row.

pub fn new_col(values: S) -> Self

Create a new RowMajorMatrix containing a single column.

pub fn as_view(&self) -> RowMajorMatrixView<'_, T>

Get a view of the matrix, i.e. a reference to the underlying data.

pub fn as_view_mut(&mut self) -> RowMajorMatrixViewMut<'_, T>
where S: BorrowMut<[T]>,

Get a mutable view of the matrix, i.e. a mutable reference to the underlying data.

pub fn copy_from<S2>(&mut self, source: &DenseMatrix<T, S2>)
where T: Copy, S: BorrowMut<[T]>, S2: DenseStorage<T>,

Copy the values from the given matrix into this matrix.

pub fn flatten_to_base<F: Field>(self) -> RowMajorMatrix<F>
where T: ExtensionField<F>,

Flatten an extension field matrix to a base field matrix.

pub fn row_slices(&self) -> impl DoubleEndedIterator<Item = &[T]>

Get an iterator over the rows of the matrix.

pub fn par_row_slices(&self) -> impl IndexedParallelIterator<Item = &[T]>
where T: Sync,

Get a parallel iterator over the rows of the matrix.

pub fn row_mut(&mut self, r: usize) -> &mut [T]
where S: BorrowMut<[T]>,

Returns a slice of the given row.

Panics

Panics if r larger than self.height().

pub fn rows_mut(&mut self) -> impl Iterator<Item = &mut [T]>
where S: BorrowMut<[T]>,

Get a mutable iterator over the rows of the matrix.

pub fn par_rows_mut<'a>( &'a mut self, ) -> impl IndexedParallelIterator<Item = &'a mut [T]>
where T: 'a + Send, S: BorrowMut<[T]>,

Get a mutable parallel iterator over the rows of the matrix.

pub fn horizontally_packed_row_mut<P>( &mut self, r: usize, ) -> (&mut [P], &mut [T])
where P: PackedValue<Value = T>, S: BorrowMut<[T]>,

Get a mutable iterator over the rows of the matrix which packs the rows into packed values.

If P::WIDTH does not divide self.width, the remainder of the row will be returned as a base slice.

pub fn scale_row(&mut self, r: usize, scale: T)
where T: Field, S: BorrowMut<[T]>,

Scale the given row by the given value.

Panics

Panics if r larger than self.height().

pub fn par_scale_row(&mut self, r: usize, scale: T)
where T: Field, S: BorrowMut<[T]>,

Scale the given row by the given value.

Performance

This function is parallelized, which may introduce some overhead compared to Self::scale_row when the width is small.

Panics

Panics if r larger than self.height().

pub fn scale(&mut self, scale: T)
where T: Field, S: BorrowMut<[T]>,

Scale the entire matrix by the given value.

pub fn split_rows( &self, r: usize, ) -> (RowMajorMatrixView<'_, T>, RowMajorMatrixView<'_, T>)

Split the matrix into two matrix views, one with the first r rows and one with the remaining rows.

Panics

Panics if r larger than self.height().

pub fn split_rows_mut( &mut self, r: usize, ) -> (RowMajorMatrixViewMut<'_, T>, RowMajorMatrixViewMut<'_, T>)
where S: BorrowMut<[T]>,

Split the matrix into two mutable matrix views, one with the first r rows and one with the remaining rows.

Panics

Panics if r larger than self.height().

pub fn par_row_chunks( &self, chunk_rows: usize, ) -> impl IndexedParallelIterator<Item = RowMajorMatrixView<'_, T>>
where T: Send,

Get an iterator over the rows of the matrix which takes chunk_rows rows at a time.

If chunk_rows does not divide the height of the matrix, the last chunk will be smaller.

pub fn par_row_chunks_exact( &self, chunk_rows: usize, ) -> impl IndexedParallelIterator<Item = RowMajorMatrixView<'_, T>>
where T: Send,

Get a parallel iterator over the rows of the matrix which takes chunk_rows rows at a time.

If chunk_rows does not divide the height of the matrix, the last chunk will be smaller.

pub fn par_row_chunks_mut( &mut self, chunk_rows: usize, ) -> impl IndexedParallelIterator<Item = RowMajorMatrixViewMut<'_, T>>
where T: Send, S: BorrowMut<[T]>,

Get a mutable iterator over the rows of the matrix which takes chunk_rows rows at a time.

If chunk_rows does not divide the height of the matrix, the last chunk will be smaller.

pub fn row_chunks_exact_mut( &mut self, chunk_rows: usize, ) -> impl Iterator<Item = RowMajorMatrixViewMut<'_, T>>
where T: Send, S: BorrowMut<[T]>,

Get a mutable iterator over the rows of the matrix which takes chunk_rows rows at a time.

If chunk_rows does not divide the height of the matrix, the last up to chunk_rows - 1 rows of the matrix will be omitted.

pub fn par_row_chunks_exact_mut( &mut self, chunk_rows: usize, ) -> impl IndexedParallelIterator<Item = RowMajorMatrixViewMut<'_, T>>
where T: Send, S: BorrowMut<[T]>,

Get a parallel mutable iterator over the rows of the matrix which takes chunk_rows rows at a time.

If chunk_rows does not divide the height of the matrix, the last up to chunk_rows - 1 rows of the matrix will be omitted.

pub fn row_pair_mut( &mut self, row_1: usize, row_2: usize, ) -> (&mut [T], &mut [T])
where S: BorrowMut<[T]>,

Get a pair of mutable slices of the given rows.

Panics

Panics if row_1 or row_2 are out of bounds or if row_1 >= row_2.

pub fn packed_row_pair_mut<P>( &mut self, row_1: usize, row_2: usize, ) -> ((&mut [P], &mut [T]), (&mut [P], &mut [T]))
where S: BorrowMut<[T]>, P: PackedValue<Value = T>,

Get a pair of mutable slices of the given rows, both packed into packed field elements.

If P:WIDTH does not divide self.width, the remainder of the row will be returned as a base slice.

Panics

Panics if row_1 or row_2 are out of bounds or if row_1 >= row_2.

pub fn bit_reversed_zero_pad(self, added_bits: usize) -> RowMajorMatrix<T>
where T: Field,

Append zeros to the “end” of the given matrix, except that the matrix is in bit-reversed order, so in actuality we’re interleaving zero rows.

impl<T: Clone + Default + Send + Sync> DenseMatrix<T>

pub fn as_cow<'a>(self) -> RowMajorMatrixCow<'a, T>

pub fn rand<R: Rng>(rng: &mut R, rows: usize, cols: usize) -> Self
where StandardUniform: Distribution<T>,

pub fn rand_nonzero<R: Rng>(rng: &mut R, rows: usize, cols: usize) -> Self
where T: Field, StandardUniform: Distribution<T>,

pub fn pad_to_height(&mut self, new_height: usize, fill: T)

pub fn pad_to_power_of_two_height(&mut self, fill: T)

Pad the matrix height to the next power of two by appending rows filled with fill.

This is commonly used in proof systems where trace matrices must have power-of-two heights.

Behavior

If the matrix is empty (height = 0), it is padded to have exactly one row of fill values.

If the height is already a power of two, the matrix is unchanged.

Otherwise, the matrix is padded to the next power of two height.

impl<T: Copy + Default + Send + Sync, V: DenseStorage<T>> DenseMatrix<T, V>

pub fn transpose(&self) -> RowMajorMatrix<T>

Return the transpose of this matrix.

pub fn transpose_into<W: DenseStorage<T> + BorrowMut<[T]>>( &self, other: &mut DenseMatrix<T, W>, )

Transpose the matrix returning the result in other without intermediate allocation.

impl<'a, T: Clone + Default + Send + Sync> DenseMatrix<T, &'a [T]>

pub fn as_cow(self) -> RowMajorMatrixCow<'a, T>