p3_matrix::row_index_mapped

Trait RowIndexMap

pub trait RowIndexMap: Send + Sync {
    // Required methods
    fn height(&self) -> usize;
    fn map_row_index(&self, r: usize) -> usize;

    // Provided method
    fn to_row_major_matrix<T: Clone + Send + Sync, Inner: Matrix<T>>(
        &self,
        inner: Inner,
    ) -> RowMajorMatrix<T> { ... }
}

A RowIndexMap remaps row indices, and can change the height.

Required Methods

fn height(&self) -> usize

fn map_row_index(&self, r: usize) -> usize

Provided Methods

fn to_row_major_matrix<T: Clone + Send + Sync, Inner: Matrix<T>>( &self, inner: Inner, ) -> RowMajorMatrix<T>

Permutations can optionally provide an optimized method to convert to dense form.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.