p3_matrix::stack

Type Alias ViewPair

pub type ViewPair<'a, T> = VerticalPair<RowMajorMatrixView<'a, T>, RowMajorMatrixView<'a, T>>;

A type alias representing a vertical composition of two row-major matrix views.

ViewPair combines two RowMajorMatrixView’s with the same element type T and lifetime 'a into a single virtual matrix stacked vertically.

Both views must have the same width; the resulting view has a height equal to the sum of the two original heights.

Aliased Type

pub struct ViewPair<'a, T> {
    pub top: DenseMatrix<T, &'a [T]>,
    pub bottom: DenseMatrix<T, &'a [T]>,
}

Fields
top: DenseMatrix<T, &'a [T]>

The top matrix in the vertical composition.

bottom: DenseMatrix<T, &'a [T]>

The bottom matrix in the vertical composition.